requirements.txt
# Core ML/AI Dependencies
torch>=2.0.0
transformers>=4.30.0
bitsandbytes>=0.41.0
datasets>=2.0.0
accelerate>=0.20.0

# Data Processing
numpy>=1.24.0
pandas>=1.5.0
ruamel.yaml>=0.17.0

# Utilities
tqdm>=4.64.0
gpustat>=1.1.0
psutil>=5.9.0

# Evaluation (Optional)
scikit-learn>=1.3.0

# Development (Optional - remove if not needed)
jupyter>=1.0.0
matplotlib>=3.5.0
seaborn>=0.12.0
